# contracts/housing-service.openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HousingService",
    "version": "1.0"
  },
  "paths": {
...
    "/api/housing-requests": {
      "post": {
        "tags": [
          "HousingRequests"
        ],
        "parameters": [
          {
            "name": "Gender",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Gender"
            }
          },
          {
            "name": "GovernorateId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "AcademicLevel",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AcademicLevel"
            }
          },
          {
            "name": "DetailedAddress",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "HasSpecialNeeds",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "IsPreviousResident",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "PreviousBuildingId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PreviousFloor",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PreviousRoomNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Notes",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "PersonalPhoto": {
                    "type": "string",
                    "format": "binary"
                  },
                  "NationalIdFront": {
                    "type": "string",
                    "format": "binary"
                  },
                  "NationalIdBack": {
                    "type": "string",
                    "format": "binary"
                  },
                  "UniversityIdFront": {
                    "type": "string",
                    "format": "binary"
                  },
                  "UniversityIdBack": {
                    "type": "string",
                    "format": "binary"
                  },
                  "DepartureReceipt": {
                    "type": "string",
                    "format": "binary"
                  },
                  "ResidencyProof": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "PersonalPhoto": {
                  "style": "form"
                },
                "NationalIdFront": {
                  "style": "form"
                },
                "NationalIdBack": {
                  "style": "form"
                },
                "UniversityIdFront": {
                  "style": "form"
                },
                "UniversityIdBack": {
                  "style": "form"
                },
                "DepartureReceipt": {
                  "style": "form"
                },
                "ResidencyProof": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "HousingRequests"
        ],
        "parameters": [
          {
            "name": "housingCycleId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "governorateId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/HousingRequestStatus"
            }
          },
          {
            "name": "admissionStatus",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AdmissionDecisionStatus"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/housing-requests/mine/{id}": {
      "get": {
        "tags": [
          "HousingRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "HousingRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Gender",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Gender"
            }
          },
          {
            "name": "GovernorateId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "AcademicLevel",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AcademicLevel"
            }
          },
          {
            "name": "DetailedAddress",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "HasSpecialNeeds",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "IsPreviousResident",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "PreviousBuildingId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PreviousFloor",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PreviousRoomNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Notes",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "PersonalPhoto": {
                    "type": "string",
                    "format": "binary"
                  },
                  "NationalIdFront": {
                    "type": "string",
                    "format": "binary"
                  },
                  "NationalIdBack": {
                    "type": "string",
                    "format": "binary"
                  },
                  "UniversityIdFront": {
                    "type": "string",
                    "format": "binary"
                  },
                  "UniversityIdBack": {
                    "type": "string",
                    "format": "binary"
                  },
                  "DepartureReceipt": {
                    "type": "string",
                    "format": "binary"
                  },
                  "ResidencyProof": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "PersonalPhoto": {
                  "style": "form"
                },
                "NationalIdFront": {
                  "style": "form"
                },
                "NationalIdBack": {
                  "style": "form"
                },
                "UniversityIdFront": {
                  "style": "form"
                },
                "UniversityIdBack": {
                  "style": "form"
                },
                "DepartureReceipt": {
                  "style": "form"
                },
                "ResidencyProof": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/housing-requests/{id}": {
      "get": {
        "tags": [
          "HousingRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/housing-requests/{id}/documents/{documentId}/review": {
      "post": {
        "tags": [
          "HousingRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDocumentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDocumentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/housing-requests/{id}/decision": {
      "post": {
        "tags": [
          "HousingRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeAdmissionDecisionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeAdmissionDecisionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MakeAdmissionDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingRequestDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcademicLevel": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "AdmissionDecisionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/AdmissionDecisionStatus"
          },
          "decisionReason": {
            "type": "string",
            "nullable": true
          },
          "decisionDate": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedBy": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AdmissionDecisionStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "DocumentReviewStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "DocumentType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Gender": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "HousingRequestDocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "documentPath": {
            "type": "string",
            "nullable": true
          },
          "reviewStatus": {
            "$ref": "#/components/schemas/DocumentReviewStatus"
          },
          "reviewNotes": {
            "type": "string",
            "nullable": true
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HousingRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "studentId": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "governorateId": {
            "type": "integer",
            "format": "int32"
          },
          "academicLevel": {
            "$ref": "#/components/schemas/AcademicLevel"
          },
          "housingCycleId": {
            "type": "integer",
            "format": "int32"
          },
          "detailedAddress": {
            "type": "string",
            "nullable": true
          },
          "hasSpecialNeeds": {
            "type": "boolean"
          },
          "isPreviousResident": {
            "type": "boolean"
          },
          "previousBuildingId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "previousFloor": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "previousRoomNumber": {
            "type": "string",
            "nullable": true
          },
          "housingGroupId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/HousingRequestStatus"
          },
          "specialNotes": {
            "type": "string",
            "nullable": true
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HousingRequestDocumentDto"
            },
            "nullable": true
          },
          "decision": {
            "$ref": "#/components/schemas/AdmissionDecisionDto"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HousingRequestDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HousingRequestDto"
            },
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "HousingRequestStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "MakeAdmissionDecisionDto": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/AdmissionDecisionStatus"
          },
          "decisionReason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReviewDocumentDto": {
        "type": "object",
        "properties": {
          "approve": {
            "type": "boolean"
          },
          "reviewNotes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}